+++ ezxml/build.rs
@@ -4,10 +4,6 @@ use std::env;
 use std::fs::File;
 use std::io::Write;
 use std::path::PathBuf;
-use std::process::Command;
-
-// use test_dir::xml_data_dir;
-use xtest::Syntax;
 
 fn main() {
     generate_readme();
@@ -33,7 +29,7 @@ fn generate_readme() {
         false, // add license
         true,  // indent headings
     )
-    .unwrap();
+        .unwrap();
 
     let mut readme = File::create("README.md").unwrap();
     readme.write_all(content.as_bytes()).unwrap();
