+++ ezxml/src/parser/mod.rs
@@ -2,14 +2,17 @@ extern crate env_logger;
 
 use std::str::Chars;
 
-use snafu::{Backtrace, GenerateBacktrace, ResultExt};
+use snafu::{Backtrace, GenerateBacktrace};
 
 pub use ds::Stack;
-use xdoc::{Declaration, Document, ElementData, Encoding, OrdMap, PIData, Version};
+use xdoc::{Declaration, Document, ElementData, Encoding, PIData, Version};
 
 use crate::error::{self, Result};
 use crate::Node;
-use crate::parser::TagStatus::OutsideTag;
+use crate::parser::pi::parse_pi;
+
+mod pi;
+mod chars;
 
 #[derive(Debug, Clone, Copy, Eq, PartialOrd, PartialEq, Hash)]
 pub struct Position {
@@ -42,12 +45,12 @@ impl Position {
 }
 
 #[derive(Debug, Clone, Eq, PartialOrd, PartialEq, Hash)]
-struct ParserState {
-    position: Position,
-    current_char: char,
-    doc_status: DocStatus,
-    tag_status: TagStatus,
-    stack: Option<Stack<crate::Node>>,
+pub(crate) struct ParserState {
+    pub(crate) position: Position,
+    pub(crate) current_char: char,
+    pub(crate) doc_status: DocStatus,
+    pub(crate) tag_status: TagStatus,
+    pub(crate) stack: Option<Stack<crate::Node>>,
 }
 
 pub fn parse_str(s: &str) -> Result<Document> {
@@ -55,7 +58,7 @@ pub fn parse_str(s: &str) -> Result<Document> {
         position: Default::default(),
         current_char: '\0',
         doc_status: DocStatus::default(),
-        tag_status: OutsideTag,
+        tag_status: TagStatus::OutsideTag,
         stack: None,
     };
 
@@ -70,8 +73,10 @@ pub fn parse_str(s: &str) -> Result<Document> {
     Ok(document)
 }
 
+// TODO - disallow dead code
+#[allow(dead_code)]
 #[derive(Debug, Clone, Copy, Eq, PartialOrd, PartialEq, Hash)]
-enum TagStatus {
+pub(crate) enum TagStatus {
     TagOpen(u64),
     InsideTag(u64),
     InsideProcessingInstruction(u64),
@@ -85,8 +90,10 @@ impl Default for TagStatus {
     }
 }
 
+// TODO - disallow dead code
+#[allow(dead_code)]
 #[derive(Debug, Clone, Copy, Eq, PartialOrd, PartialEq, Hash)]
-enum DocStatus {
+pub(crate) enum DocStatus {
     BeforeDeclaration,
     AfterDeclaration,
     BeforeRoot,
@@ -100,13 +107,6 @@ impl Default for DocStatus {
     }
 }
 
-fn is_space_or_alpha(c: char) -> bool {
-    c.is_alphabetic() || c.is_ascii_whitespace()
-}
-
-fn is_pi_indicator(c: char) -> bool {
-    c == '?'
-}
 
 fn parse_document(iter: &mut Chars, state: &mut ParserState, document: &mut Document) -> Result<()> {
     loop {
@@ -146,245 +146,8 @@ fn parse_document(iter: &mut Chars, state: &mut ParserState, document: &mut Docu
     Ok(())
 }
 
-#[derive(PartialEq)]
-enum PIStatus {
-    BeforeTarget,
-    InsideTarget,
-    AfterTarget,
-    InsideKey,
-    AfterKey,
-    Equals,
-    AfterEquals,
-    ValOpenQuote,
-    InsideVal,
-    ValCloseQuote,
-    AfterVal,
-    QuestionMark,
-    Close,
-}
 
-struct PIProcessor {
-    status: PIStatus,
-    key_buffer: String,
-    value_buffer: String,
-    pi_data: PIData,
-}
-
-impl PIProcessor {
-    fn new() -> Self {
-        Self {
-            status: PIStatus::BeforeTarget,
-            key_buffer: "".to_string(),
-            value_buffer: "".to_string(),
-            pi_data: PIData::default(),
-        }
-    }
-
-    /// Takes the current strings from `key_buffer` and `value_buffer` and adds them to the
-    /// `instructions`. Clears these buffers to begin processing the next key/value pair.
-    fn take_buffers(&mut self) -> Result<()> {
-        if self.key_buffer.is_empty() {
-            // TODO - better error
-            return Err(error::Error::Bug { message: "Empty key - this is a bug and should have been detected sooner.".to_string() });
-        }
-        if let Some(_) = self.pi_data.instructions.mut_map().insert(self.key_buffer.clone(), self.value_buffer.clone()) {
-            // TODO - better error
-            return Err(error::Error::Bug { message: "Duplicate key".to_string() });
-        }
-        self.key_buffer.clear();
-        self.value_buffer.clear();
-        Ok(())
-    }
-}
-
-fn parse_pi(iter: &mut Chars, state: &mut ParserState) -> Result<PIData> {
-    let mut processor = PIProcessor::new();
-    loop {
-        if let Err(e) = take_processing_instruction_char(iter, state, &mut processor) {
-            return Err(e);
-        }
-        if processor.status == PIStatus::Close {
-            break;
-        }
-
-        if !advance_parser(iter, state) {
-            return Err(error::Error::Parse {
-                position: state.position,
-                backtrace: Backtrace::generate(),
-            });
-        }
-    }
-
-    Ok(processor.pi_data)
-}
-
-// for valid name start char ranges
-const U_00C0: char = '\u{00C0}';
-const U_00D6: char = '\u{00D6}';
-const U_00D8: char = '\u{00D8}';
-const U_00F6: char = '\u{00F6}';
-const U_00F8: char = '\u{00F8}';
-const U_02FF: char = '\u{02FF}';
-const U_0370: char = '\u{0370}';
-const U_037D: char = '\u{037D}';
-const U_037F: char = '\u{037F}';
-const U_1FFF: char = '\u{1FFF}';
-const U_200C: char = '\u{200C}';
-const U_200D: char = '\u{200D}';
-const U_2070: char = '\u{2070}';
-const U_218F: char = '\u{218F}';
-const U_2C00: char = '\u{2C00}';
-const U_2FEF: char = '\u{2FEF}';
-const U_3001: char = '\u{3001}';
-const U_D7FF: char = '\u{D7FF}';
-const U_F900: char = '\u{F900}';
-const U_FDCF: char = '\u{FDCF}';
-const U_FDF0: char = '\u{FDF0}';
-const U_FFFD: char = '\u{FFFD}';
-const U_10000: char = '\u{10000}';
-const U_EFFFF: char = '\u{EFFFF}';
-
-// for valid name char ranges
-const U_00B7: char = '\u{00B7}';
-const U_0300: char = '\u{0300}';
-const U_036F: char = '\u{036F}';
-const U_203F: char = '\u{203F}';
-const U_2040: char = '\u{2040}';
-
-
-fn take_processing_instruction_char(iter: &mut Chars, state: &mut ParserState, processor: &mut PIProcessor) -> Result<()> {
-    let ch = state.current_char;
-    println!("{}", ch);
-    match processor.status {
-        PIStatus::BeforeTarget => {
-            if !is_name_start_char(state.current_char) {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            } else {
-                processor.pi_data.target.push(state.current_char);
-                processor.status = PIStatus::InsideTarget;
-            }
-        }
-        PIStatus::InsideTarget => {
-            if state.current_char.is_ascii_whitespace() {
-                processor.status = PIStatus::AfterTarget;
-            } else if !is_name_char(state.current_char) {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            } else {
-                processor.pi_data.target.push(state.current_char);
-            }
-        }
-        PIStatus::AfterTarget => {
-            if is_name_start_char(state.current_char) {
-                processor.key_buffer.push(state.current_char);
-                processor.status = PIStatus::InsideKey;
-            } else if !state.current_char.is_ascii_whitespace() {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            }
-        }
-        PIStatus::InsideKey => {
-            if is_name_char(state.current_char) {
-                processor.key_buffer.push(state.current_char);
-                processor.status = PIStatus::InsideKey;
-            } else if state.current_char == '=' {
-                processor.status = PIStatus::Equals;
-            } else if state.current_char.is_ascii_whitespace() {
-                processor.status = PIStatus::AfterKey;
-            } else {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            }
-        }
-        PIStatus::AfterKey => {
-            if state.current_char == '=' {
-                processor.status = PIStatus::Equals;
-            } else if !state.current_char.is_ascii_whitespace() {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            }
-        }
-        PIStatus::Equals | PIStatus::AfterEquals => {
-            if state.current_char == '"' {
-                processor.status = PIStatus::ValOpenQuote;
-            } else if state.current_char.is_ascii_whitespace() {
-                processor.status = PIStatus::AfterEquals;
-            } else {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            }
-        }
-        PIStatus::ValOpenQuote | PIStatus::InsideVal => {
-            if state.current_char == '"' {
-                if let Err(_) = processor.take_buffers() {
-                    return Err(error::Error::Parse {
-                        position: state.position,
-                        backtrace: Backtrace::generate(),
-                    });
-                }
-                processor.status = PIStatus::ValCloseQuote;
-            } else {
-                // TODO - handle escape sequences
-                processor.value_buffer.push(state.current_char);
-                processor.status = PIStatus::InsideVal;
-            }
-        }
-        PIStatus::ValCloseQuote => {
-            if state.current_char.is_ascii_whitespace() {
-                processor.status = PIStatus::AfterVal;
-            } else if state.current_char == '?' {
-                processor.status = PIStatus::QuestionMark;
-            } else {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            }
-        }
-        PIStatus::AfterVal => {
-            if state.current_char.is_ascii_whitespace() {
-                processor.status = PIStatus::AfterVal;
-            } else if state.current_char == '?' {
-                processor.status = PIStatus::QuestionMark;
-            } else if is_name_start_char(state.current_char) {
-                processor.key_buffer.push(state.current_char);
-                processor.status = PIStatus::InsideKey;
-            } else {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            }
-        }
-        PIStatus::QuestionMark => {
-            if state.current_char == '>' {
-                processor.status = PIStatus::Close;
-            } else {
-                return Err(error::Error::Parse {
-                    position: state.position,
-                    backtrace: Backtrace::generate(),
-                });
-            }
-        }
-        PIStatus::Close => { /* done */ }
-    }
-    Ok(())
-}
-
-fn advance_parser(iter: &mut Chars<'_>, state: &mut ParserState) -> bool {
+pub(crate) fn advance_parser(iter: &mut Chars<'_>, state: &mut ParserState) -> bool {
     let option_char = iter.next();
     match option_char {
         Some(c) => {
@@ -396,7 +159,7 @@ fn advance_parser(iter: &mut Chars<'_>, state: &mut ParserState) -> bool {
     }
 }
 
-fn advance_parser_or_die(iter: &mut Chars<'_>, state: &mut ParserState) -> Result<()> {
+pub(crate) fn advance_parser_or_die(iter: &mut Chars<'_>, state: &mut ParserState) -> Result<()> {
     if advance_parser(iter, state) {
         Ok(())
     } else {
@@ -407,56 +170,9 @@ fn advance_parser_or_die(iter: &mut Chars<'_>, state: &mut ParserState) -> Resul
     }
 }
 
-fn is_name_start_char(c: char) -> bool {
-    // TODO oops make sure its the same as 1.1 https://www.w3.org/TR/2006/REC-xml11-20060816/
-    // https://www.w3.org/TR/2008/REC-xml-20081126/#NT-NameStartChar
-    // [4]   	NameStartChar	   ::=   	":" | [A-Z] | "_" | [a-z] | [#xC0-#xD6] | [#xD8-#xF6] |
-    // [#xF8-#x2FF] | [#x370-#x37D] | [#x37F-#x1FFF] | [#x200C-#x200D] | [#x2070-#x218F] |
-    // [#x2C00-#x2FEF] | [#x3001-#xD7FF] | [#xF900-#xFDCF] | [#xFDF0-#xFFFD] | [#x10000-#xEFFFF]
-    // let x = c as u64;
-    match c {
-        'A'..='Z' => true,
-        'a'..='z' => true,
-        ':' => true,
-        '_' => true,
-        U_00C0..=U_00D6 => true,
-        U_00D8..=U_00F6 => true,
-        U_00F8..=U_02FF => true,
-        U_0370..=U_037D => true,
-        U_037F..=U_1FFF => true,
-        U_200C..=U_200D => true,
-        U_2070..=U_218F => true,
-        U_2C00..=U_2FEF => true,
-        U_3001..=U_D7FF => true,
-        U_F900..=U_FDCF => true,
-        U_FDF0..=U_FFFD => true,
-        U_10000..=U_EFFFF => true,
-        _ => false,
-    }
-}
-
-fn is_name_char(c: char) -> bool {
-    // TODO oops make sure its the same as 1.1 https://www.w3.org/TR/2006/REC-xml11-20060816/
-    // https://www.w3.org/TR/2008/REC-xml-20081126/#NT-NameChar
-    // [4a] NameChar ::= NameStartChar | "-" | "." | [0-9] | #xB7 | [#x0300-#x036F] | [#x203F-#x2040]
-    if is_name_start_char(c) {
-        return true;
-    }
-    match c {
-        U_00B7 => true,
-        U_0300..=U_036F => true,
-        U_203F..=U_2040 => true,
-        '0'..='9' => true,
-        '-' => true,
-        '.' => true,
-        _ => false,
-    }
-}
-
-
 fn parse_declaration(pi_data: &PIData) -> Result<Declaration> {
     let mut declaration = Declaration::default();
-    if pi_data.target != "xml".to_string() {
+    if pi_data.target != "xml" {
         return Err(error::Error::Bug { message: "TODO - better message".to_string() });
     }
     if pi_data.instructions.map().len() > 2 {
@@ -492,7 +208,7 @@ fn state_must_be_before_declaration(state: &ParserState) -> Result<()> {
     }
 }
 
-fn peek_or_die(iter: &mut Chars) -> Result<char> {
+pub(crate) fn peek_or_die(iter: &mut Chars) -> Result<char> {
     let mut peekable = iter.peekable();
     let opt = peekable.peek();
     match opt {
